made dataType optional, altered tests, added helper file
